requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2-binary
alembic
python-dotenv
passlib
redis
httpx
jwt
PyJWT==1.7.1
asyncpg
greenlet
aio-pika
aiosmtplib
jinja2
python-multipart
pillow